+++ releasenotes.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Version 0.9.0
+### New Features
+- Lorem.Paragraph(int sentenceCount = 3, bool supplemental = false, int randomSentencesToAdd = 0): Optional parameter and default values added.
+
 ## Version 0.8.0
 ### New Features
 - Lorem.Sentence(int wordCount = 4, bool supplemental = false, int randomWordsToAdd = 6): Optional parameters and default values added.
@@ -18,13 +22,13 @@
 - Lorem.Multibyte()
 
 ## Version 0.6.0
+### Breaking Changes
+- Removed Lorem.GetFirstWord(). If you need to generate a defined word, use `RandomNumber.ResetSeed(int)` and then `Lorem.Word()` instead. We believe the impact of this change is minimal as the method was used for testing only. It does not exist in the Roby gem.
+
 ### New Features
 - Company.Name()
 - Lorem.Word()
 - Lorem.Words(int count = 3, bool supplemental = false): Optional parameters and default values added.
-
-### Breaking Changes
-- Removed Lorem.GetFirstWord(). If you need to generate a defined word, use `RandomNumber.ResetSeed(int)` and then `Lorem.Word()` instead. We believe the impact of this change is minimal as the method was used for testing only. It does not exist in the Roby gem.
 
 ## Version 0.5.0
 ### New Features